# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bloom Together API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bloomspaces/{id}/members:
    get:
      operationId: getBloomSpaceMembers
      tags: [bloomspaces]
      summary: Get members of a Bloom Space
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Members list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

  # EVENTS
...
  /memories:
    get:
      operationId: listMemories
      tags: [memories]
      summary: List memories for current Bloom Space
      parameters:
        - name: view
          in: query
          schema:
            type: string
            enum: [grid, timeline, masonry]
      responses:
        "200":
          description: Memories list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Memory"
    post:
      operationId: createMemory
      tags: [memories]
      summary: Create a new memory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryInput"
      responses:
        "201":
          description: Memory created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
...
  /memories/search:
    get:
      operationId: searchMemories
      tags: [memories]
      summary: Search memories by title, location, description, or favorite moments
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Memory"
  /memories/{id}:
    get:
      operationId: getMemory
      tags: [memories]
      summary: Get a memory by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Memory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
    patch:
      operationId: updateMemory
      tags: [memories]
      summary: Update a memory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryUpdate"
      responses:
        "200":
          description: Updated memory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
    delete:
      operationId: deleteMemory
      tags: [memories]
      summary: Delete a memory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /memories/{id}/photos:
    post:
      operationId: addMemoryPhoto
      tags: [memories]
      summary: Add a photo to a memory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryPhotoInput"
      responses:
        "201":
          description: Photo added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryPhoto"
  /memories/{memoryId}/photos/{photoId}:
    patch:
      operationId: updateMemoryPhoto
      tags: [memories]
      summary: Update photo caption or sort order
      parameters:
        - name: memoryId
          in: path
          required: true
          schema:
            type: integer
        - name: photoId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryPhotoUpdate"
      responses:
        "200":
          description: Updated photo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryPhoto"
    delete:
      operationId: deleteMemoryPhoto
      tags: [memories]
      summary: Remove a photo from a memory
      parameters:
        - name: memoryId
          in: path
          required: true
          schema:
            type: integer
        - name: photoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /memories/{id}/comments:
    get:
      operationId: listMemoryComments
      tags: [memories]
      summary: List comments on a memory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemoryComment"
    post:
      operationId: addMemoryComment
      tags: [memories]
      summary: Add a comment to a memory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryCommentInput"
      responses:
        "201":
          description: Comment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryComment"
  /memories/{id}/comments/{commentId}:
    delete:
      operationId: deleteMemoryComment
      tags: [memories]
      summary: Delete a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /memories/{id}/reactions:
    post:
      operationId: toggleMemoryReaction
      tags: [memories]
      summary: Toggle a reaction on a memory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryReactionInput"
      responses:
        "200":
          description: Reactions updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemoryReaction"

  # NOTES
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        favoriteFlower:
          type: ["string", "null"]
        favoriteActivity:
          type: ["string", "null"]
        createdAt:
          type: string

    # BLOOM SPACE SCHEMAS
...
    MemoryInput:
      type: object
      required: [title, date]
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        date:
          type: string
        location:
          type: ["string", "null"]
        activityType:
          type: ["string", "null"]
        favoriteMoment:
          type: ["string", "null"]
        partnerFavoriteMoment:
          type: ["string", "null"]
        moodRating:
          type: ["integer", "null"]
        coverImageUrl:
          type: ["string", "null"]
    MemoryUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        date:
          type: string
        location:
          type: ["string", "null"]
        activityType:
          type: ["string", "null"]
        favoriteMoment:
          type: ["string", "null"]
        partnerFavoriteMoment:
          type: ["string", "null"]
        moodRating:
          type: ["integer", "null"]
        coverImageUrl:
          type: ["string", "null"]
    Memory:
      type: object
      required: [id, bloomSpaceId, title, date, createdBy]
      properties:
        id:
          type: integer
        bloomSpaceId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        date:
          type: string
        location:
          type: ["string", "null"]
        activityType:
          type: ["string", "null"]
        favoriteMoment:
          type: ["string", "null"]
        partnerFavoriteMoment:
          type: ["string", "null"]
        moodRating:
          type: ["integer", "null"]
        coverImageUrl:
          type: ["string", "null"]
        createdBy:
          type: integer
        createdAt:
          type: string
        photos:
          type: array
          items:
            $ref: "#/components/schemas/MemoryPhoto"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/MemoryComment"
        reactions:
          type: array
          items:
            $ref: "#/components/schemas/MemoryReaction"
        creator:
          $ref: "#/components/schemas/User"
        partner:
          $ref: "#/components/schemas/User"
    MemoryPhotoInput:
      type: object
      required: [imageUrl]
      properties:
        imageUrl:
          type: string
        caption:
          type: ["string", "null"]
        sortOrder:
          type: ["integer", "null"]
    MemoryPhotoUpdate:
      type: object
      properties:
        caption:
          type: ["string", "null"]
        sortOrder:
          type: ["integer", "null"]
    MemoryPhoto:
      type: object
      required: [id, memoryId, imageUrl]
      properties:
        id:
          type: integer
        memoryId:
          type: integer
        imageUrl:
          type: string
        caption:
          type: ["string", "null"]
        sortOrder:
          type: ["integer", "null"]
    MemoryCommentInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
    MemoryComment:
      type: object
      required: [id, memoryId, userId, content, createdAt]
      properties:
        id:
          type: integer
        memoryId:
          type: integer
        userId:
          type: integer
        content:
          type: string
        createdAt:
          type: string
        author:
          $ref: "#/components/schemas/User"
    MemoryReactionInput:
      type: object
      required: [reactionType]
      properties:
        reactionType:
          type: string
          enum: [rose, peony, tulip, daisy, heart]
    MemoryReaction:
      type: object
      required: [id, memoryId, userId, reactionType]
      properties:
        id:
          type: integer
        memoryId:
          type: integer
        userId:
          type: integer
        reactionType:
          type: string

    # NOTE SCHEMAS
...